Rename project clojure_like_threading_macr to conamon
