+++ Gemfile
@@ -2,5 +2,5 @@ source 'https://rubygems.org'
 
 gem 'rspec'
 
-# Specify your gem's dependencies in clojure_like_threading_macro.gemspec
+# Specify your gem's dependencies in conamon.gemspec
 gemspec
