--- lib/clojure_like_threading_macro/version.rb
@@ -1,3 +1,0 @@
-module ClojureLikeThreadingMacro
-  VERSION = "0.1.0"
-end
