+++ lib/conamon/bind_marker.rb
