+++ lib/conamon/version.rb
@@ -1,0 +1,3 @@
+module Conamon
+  VERSION = "0.1.0"
+end
